+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggmagnify (development version)
 
+* New `proj.fill` argument to fill area between the projection lines.
+
 
 # ggmagnify 0.3.0